The Reddit Backend/API add user checking

+++ week-04/day-04/The-Reddit-Backend/db.js
@@ -73,7 +73,6 @@ function votePosts(_id, voteType, callback) {
 
 function updateData(db, whereStr, updateStr) {
   var collection = db.collection('posts');
-  updateStr.timestamp = Date.parse(new Date());
   collection.update(whereStr,updateStr, function(err) {
     if(err){
       console.log('Error:'+ err);
@@ -83,13 +82,17 @@ function updateData(db, whereStr, updateStr) {
 }
 
 // delete posts
-function deletePosts(_id, callback) {
+function deletePosts(_id, username, callback) {
   var whereStr = {'_id': ObjectId(_id)};
   var fieldStr = {};
   MongoClient.connect(url, function (err, db) {
     selectData(db, whereStr, fieldStr, function(result) {
-      deleteData(db, whereStr);
-      callback(result);
+      if (result[0].owner === username || !result[0].owner) {
+        deleteData(db, whereStr);
+        callback(result[0]);
+      }else {
+        callback({'error': 'You have no right to delete this post!'});
+      }
       db.close();
     });
   });
@@ -97,7 +100,6 @@ function deletePosts(_id, callback) {
 
 function deleteData(db, whereStr) {
   var collection = db.collection('posts');
-  console.log(whereStr);
   collection.remove(whereStr, function(err) {
     if(err){
       console.log('Error:'+ err);
@@ -114,10 +116,15 @@ function modifyPosts(obj, callback) {
   MongoClient.connect(url, function (err, db) {
     selectData(db, whereStr, fieldStr, function(result) {
       updateStr = result[0];
-      updateStr.title = obj.title;
-      updateStr.href = obj.href;
-      updateData(db, whereStr, updateStr)
-      callback(updateStr);
+      if (updateStr.owner === obj.owner || !updateStr.owner){
+        updateStr.title = obj.title;
+        updateStr.href = obj.href;
+        updateStr.timestamp = Date.parse(new Date());
+        updateData(db, whereStr, updateStr)
+        callback(updateStr);
+      }else {
+        callback({'error': 'You have no right to modify this post!'});
+      }
       db.close();
     });
   });
